basic controller

+++ src/main/java/com/cg/sprint1_onlineplantnursery/service/IPlanterServiceImpl.java
@@ -1,6 +1,9 @@
 package com.cg.sprint1_onlineplantnursery.service;
 
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.cg.sprint1_onlineplantnursery.entity.Planter;
@@ -14,15 +17,30 @@ public class IPlanterServiceImpl implements IPlanterService {
 
 	@Override
 	public Planter addPlanter(Planter planter) {
-		return iplanterRepo.save(planter);
+		Optional<Planter> op = iplanterRepo.findById(planter.getPlanterId());
+		if (op.isPresent()) {
+			Planter p = op.get();
+			p.setPlanterStock(p.getPlanterStock()+1);
+			return iplanterRepo.save(p);
+		}else {
+			return iplanterRepo.save(planter);
+		}
+		
+		
+		
 	}
 
 	@Override
 	public Planter deletePlanter(Planter planter) {
 		Planter p = iplanterRepo.findById(planter.getPlanterId()).get();
-		if (p != null)
-			iplanterRepo.delete(p);
-		return planter;
+		if (p != null) {
+			p.setPlanterStock(p.getPlanterStock()-1);
+			if (p.getPlanterStock() < 1)
+				iplanterRepo.delete(p);
+			else
+				iplanterRepo.save(p);
+		}
+		return iplanterRepo.findById(planter.getPlanterId()).get();
 	}
 
 	@Override
@@ -38,28 +56,30 @@ public class IPlanterServiceImpl implements IPlanterService {
 
 	@Override
 	public List<Planter> viewAllPlanters(double minCost, double maxCost) {
-		List<Planter> requiredPlanters = iplanterRepo.findAll();
-		for (Planter planter: requiredPlanters) {
-			if (planter.getPlanterCost() < minCost || planter.getPlanterCost() > maxCost)
-				requiredPlanters.remove(planter);
-		}
+		List<Planter> allPlanters = iplanterRepo.findAll();
+		List<Planter> requiredPlanters = allPlanters.stream().filter((p) -> p.getPlanterCost() >minCost && p.getPlanterCost() < maxCost).collect(Collectors.toList());
+		/*
+		 * for (Planter planter: requiredPlanters) { if (planter.getPlanterCost() <
+		 * minCost || planter.getPlanterCost() > maxCost)
+		 * requiredPlanters.remove(planter); }
+		 */
 		return requiredPlanters;
 	}
 
 	@Override
-	public Planter viewPlanter(String planterShape) {
-		List<Planter> requiredPlanters = iplanterRepo.findAll();
-		for (Planter planter: requiredPlanters) {
-			if (planter.getPlanterShape().equals(planterShape))
-				return planter;
-		}
-		return null;
+	public List<Planter> viewPlanter(String planterShape) {
+		List<Planter> allPlanters = iplanterRepo.findAll();
+		List<Planter> requiredPlanters = allPlanters.stream().filter((p) -> p.getPlanterShape().equals(planterShape)).collect(Collectors.toList());
+		return requiredPlanters;
 			
 	}
 
 	@Override
 	public Planter updatePlanter(Planter planter) {
-		return iplanterRepo.save(planter);
+		Planter p = iplanterRepo.findById(planter.getPlanterId()).get();
+		p.setPlanterColor("Purple"); //Hard coded for now
+		iplanterRepo.save(p);
+		return iplanterRepo.findById(planter.getPlanterId()).get();
 	}
 	
 }
